Fix some field types

+++ priv/repo/migrations/20200607212642_create_users.exs
@@ -3,7 +3,7 @@ defmodule PolicrMini.Repo.Migrations.CreateUsers do
 
   def change do
     create table(:users, primary_key: false) do
-      add :id, :integer, comment: "用户编号", primary_key: true
+      add :id, :bigint, comment: "用户编号", primary_key: true
       add :first_name, :string, comment: "姓"
       add :last_name, :string, comment: "名"
       add :username, :string, comment: "用户名"
